+++ apps/web-demo/src/App.tsx
@@ -1,42 +1,67 @@
-import { Login, CreateWallet } from "ui";
-import { useState } from "react";
-import { NftService, Nft } from "@liquality/wallet-sdk";
+// import { Login, CreateWallet } from "ui";
+// import { useState } from "react";
+import { NftService, setup } from "@liquality/wallet-sdk";
 
-const verifierMap: Record<string, any> = {
-  google: {
-    name: "Google",
-    typeOfLogin: "google",
-    clientId:
-      "852640103435-0qhvrgpkm66c9hu0co6edkhao3hrjlv3.apps.googleusercontent.com",
-    verifier: "liquality-google-testnet",
-  },
-};
+// const verifierMap: Record<string, any> = {
+//   google: {
+//     name: "Google",
+//     typeOfLogin: "google",
+//     clientId:
+//       "852640103435-0qhvrgpkm66c9hu0co6edkhao3hrjlv3.apps.googleusercontent.com",
+//     verifier: "liquality-google-testnet",
+//   },
+// };
 
 // 1. Setup Service Provider
-const directParams = {
-  baseUrl: `http://localhost:3005/serviceworker`,
-  enableLogging: true,
-  networkUrl: "https://goerli.infura.io/v3/a8684b771e9e4997a567bbd7189e0b27",
-  network: "testnet" as any,
-};
+// const directParams = {
+//   baseUrl: `http://localhost:3005/serviceworker`,
+//   enableLogging: true,
+//   networkUrl: "https://goerli.infura.io/v3/a8684b771e9e4997a567bbd7189e0b27",
+//   network: "testnet" as any,
+// };
 
 export default function Web() {
-  const [address, setAddress] = useState<string>();
-  const [nfts, setNfts] = useState<Nft[] | null>([]);
+
+  setup({
+    alchemyApiKey:'',
+    etherscanApiKey:"-",
+    infuraProjectId:"-",
+    pocketNetworkApplicationID:"-",
+    quorum:1,
+    slowGasPriceMultiplier:1,
+    averageGasPriceMultiplier:1.5,
+    fastGasPriceMultiplier:2,
+    gasLimitMargin:2000,
+    gelatoApiKey:"",
+  });
+
+  // const [address, setAddress] = useState<string>();
+  /*  const [nfts, setNfts] = useState<Nft[] | null>([]);
 
   async function updateNfts() {
     if (!address) throw new Error("set address first");
-    const nfts = await NftService.getNfts(address, 1);
+    //const nfts = await NftService.getNfts(address, 1);
     console.log(nfts, "NFTS in my addr");
     setNfts(nfts);
+  } */
+
+  async function updateNfts() {
+    // if (!address) throw new Error("set address first");
+    // alert(window.ethereum.address)
+    // const hash = await NftService.createERC721Collection({tokenName: "gaslessNft", tokenSymbol:"gnft"}, 80001,"", true);
+        const hash = await NftService.mintERC721Token({contractAddress: "0x276d843c8c7f3aa6518b6ba119d92c6262dd3577", recipient:"0x97542289b1453eb8e9c0f4af562ef7eb354db75c", uri:"spark"}, 80001,"<private key>", true);
+
+    console.log("hash => ",hash);
+    // console.log(nfts, "NFTS in my addr");
+    // setNfts(nfts);
   }
 
   return (
     <div>
-      <Login directParams={directParams} verifierMap={verifierMap} />
-      <CreateWallet directParams={directParams} verifierMap={verifierMap} />
+      {/* <Login directParams={directParams} verifierMap={verifierMap} />
+      <CreateWallet directParams={directParams} verifierMap={verifierMap} /> */}
       <hr />
-      <div>
+      {/*  <div>
         NFT address:
         <input
           type="text"
@@ -55,7 +80,8 @@ export default function Web() {
             <div>{nft.contract.name}</div>
           ))}
         </div>
-      )}
+      )} */}
+      <button onClick={() => updateNfts()}>Get NFTS</button>
     </div>
   );
 }