+++ packages/sdk/src/account/account.service.ts
@@ -1,0 +1,11 @@
+import { ethers } from "ethers";
+import { getChainProvider } from "../factory/chain-provider";
+
+
+export abstract class AccountService {
+
+    public static async getBalance(address: string, chainID: number)  {
+        let balance = await (await getChainProvider(chainID)).getBalance(address)
+        return balance.toString()
+    }
+}