+++ packages/sdk/src/nft/nft.service.ts
@@ -11,16 +11,22 @@ import {
 } from "./types";
 import { AddressZero } from "@ethersproject/constants";
 import { NftProvider } from "./nft.provider";
-import { ethers, Wallet } from "ethers";
-import {
+import { ethers } from "ethers";
+import { ExternalProvider } from "@ethersproject/providers";
+import { 
   LiqERC1155,
+  LiqERC1155Meta__factory,
   LiqERC1155__factory,
   LiqERC721,
+  LiqERC721Meta__factory,
   LiqERC721__factory,
 } from "../../typechain-types";
 import { TransactionService } from "../transaction/transaction.service";
 import { getChainProvider } from "../factory/chain-provider";
-
+import { getWallet } from "../common/utils";
+import { GELATO_RELAY_ADDRESS, GELATO_SUPPORTED_NETWORKS } from "../common/constants";
+import { Wallet } from "alchemy-sdk";
+import { Gelato } from "../gasless-providers/gelato";
 export abstract class NftService {
   private static cache: Record<string, NftInfo> = {};
 
@@ -32,17 +38,25 @@ export abstract class NftService {
     return NftProvider.getNfts(owner, chainID);
   }
 
+  // Gets all Nfts minted from a contract. 
+  // use the pageKey( returned with each response ) to request the next page of nfts
+  // pageSize is 100 by default
+  public static async getNftsForContract(contractAddress: string, chainID: number, options?: {pageKey?: string, pageSize?: number}){
+      return NftProvider.getNftsForContract(contractAddress, chainID, options);
+  }
+
   public static async transferNft(
     transferRequest: TransferRequest,
     chainId: number,
-    pk: string
+    pkOrProvider: string | ExternalProvider,
+    isGasless: boolean
   ): Promise<string> {
     const { contractAddress, receiver, tokenIDs, amounts } =
       transferRequest;
     const { schema, contract } = await this.cacheGet(contractAddress, chainId);
 
-    const wallet = new Wallet(pk, getChainProvider(chainId));
-    const owner = wallet.address;
+    const wallet = await getWallet(pkOrProvider, chainId);
+    const owner = await wallet.getAddress();
     let tx: PopulatedTransaction;
     const data = "0x";
 
@@ -87,6 +101,8 @@ export abstract class NftService {
       }
     }
 
+    if(isGasless) return Gelato.sendTx(chainId,contractAddress,owner,tx.data!,pkOrProvider);
+
     const preparedTx = await TransactionService.prepareTransaction(
       {
         ...tx,
@@ -118,7 +134,7 @@ export abstract class NftService {
         nftType == NftType.ERC1155 ? LiqERC1155__factory : LiqERC721__factory;
       NftService.cache[contractAddress] = {
         contract: contractFactory
-          .connect(AddressZero, getChainProvider(chainID))
+          .connect(AddressZero, await getChainProvider(chainID))
           .attach(contractAddress),
         schema: nftType,
       };
@@ -132,30 +148,63 @@ export abstract class NftService {
   public static async createERC1155Collection(
     { uri }: CreateERC1155CollectionRequest,
     chainId: number,
-    pk: string
+    pkOrProvider: string | ExternalProvider,
+    isGaslessCompliant: boolean
   ): Promise<string> {
-    const contractFactory = new ethers.ContractFactory(
-      LiqERC1155__factory.abi,
-      LiqERC1155__factory.bytecode,
-      new Wallet(pk, getChainProvider(chainId))
-    );
+    const wallet = await getWallet(pkOrProvider, chainId);
+    const owner = await wallet.getAddress();
 
-    return (await contractFactory.deploy(
-      uri,
-    )).deployTransaction.hash;
+    let contractFactory;
+    let args = [uri];
+
+    if(isGaslessCompliant) {
+      if(!GELATO_SUPPORTED_NETWORKS.includes(chainId)) throw("Cannot deploy gasless compliant contract on the chain. not supported by us now");
+      contractFactory = new ethers.ContractFactory(
+        LiqERC1155Meta__factory.abi,
+        LiqERC1155Meta__factory.bytecode,
+        wallet
+      );
+      args.push(GELATO_RELAY_ADDRESS);
+
+    }else{
+      contractFactory = new ethers.ContractFactory(
+        LiqERC1155__factory.abi,
+        LiqERC1155__factory.bytecode,
+        wallet
+      );
+    }
+
+    const tx = contractFactory.getDeployTransaction(...args);
+
+    const preparedTx = await TransactionService.prepareTransaction(
+      {
+        from: owner,
+        to: undefined,
+        data: tx.data?.toString(),
+        chainId,
+      },
+      chainId
+    );
+    return (
+      await (await getWallet(pkOrProvider, chainId)).sendTransaction(
+        preparedTx
+      )
+    ).hash;
+
   }
 
   public static async mintERC1155Token(
     { contractAddress, recipient, id, amount }: MintERC1155Request,
     chainId: number,
-    pk: string
+    pkOrProvider: string | ExternalProvider,
+    isGasless: boolean
   ): Promise<string> {
     const contract = LiqERC1155__factory.connect(
       AddressZero,
-      getChainProvider(chainId)
+      await getChainProvider(chainId)
     ).attach(contractAddress);
-    const wallet = new Wallet(pk, getChainProvider(chainId));
-    const owner = wallet.address;
+    const wallet = await getWallet(pkOrProvider, chainId);
+    const owner = await wallet.getAddress();
 
     const data = "0x";
     const tx = await contract.populateTransaction.mint(
@@ -164,6 +213,8 @@ export abstract class NftService {
       amount,
       data
     );
+
+    if(isGasless) return Gelato.sendTx(chainId,contractAddress,owner,tx.data!,pkOrProvider)
 
     const preparedTx = await TransactionService.prepareTransaction(
       {
@@ -184,33 +235,68 @@ export abstract class NftService {
   public static async createERC721Collection(
     { tokenName, tokenSymbol }: CreateERC721CollectionRequest,
     chainId: number,
-    pk: string
+    pkOrProvider: string | ExternalProvider,
+    isGaslessCompliant: boolean
   ): Promise<string> {
-    const contractFactory = new ethers.ContractFactory(
-      LiqERC721__factory.abi,
-      LiqERC721__factory.bytecode,
-      new Wallet(pk, getChainProvider(chainId))
-    );
+    const wallet = await getWallet(pkOrProvider, chainId);
+    const owner = await wallet.getAddress();
 
-    return (await contractFactory.deploy(
-      tokenName,
-      tokenSymbol
-    )).deployTransaction.hash;
+    let contractFactory;
+    let args = [tokenName, tokenSymbol];
+    if(isGaslessCompliant) {
+      if(!GELATO_SUPPORTED_NETWORKS.includes(chainId)) throw("Cannot deploy gasless compliant contract on the chain. not supported by us now");
+      contractFactory = new ethers.ContractFactory(
+        LiqERC721Meta__factory.abi,
+        LiqERC721Meta__factory.bytecode,
+        wallet
+      );
+      args.push(GELATO_RELAY_ADDRESS);
+
+
+    }else{
+      contractFactory = new ethers.ContractFactory(
+        LiqERC721__factory.abi,
+        LiqERC721__factory.bytecode,
+        wallet
+      );
+    }
+
+    const tx = contractFactory.getDeployTransaction(...args);
+
+    const preparedTx = await TransactionService.prepareTransaction(
+      {
+        from: owner,
+        to: undefined,
+        data: tx.data?.toString(),
+        chainId,
+      },
+      chainId
+    );
+    return (
+      await (await getWallet(pkOrProvider, chainId)).sendTransaction(
+        preparedTx
+      )
+    ).hash;
+
   }
 
   public static async mintERC721Token(
     { contractAddress, recipient, uri }: MintERC721Request,
     chainId: number,
-    pk: string
+    pkOrProvider: string | ExternalProvider,
+    isGasless: boolean
   ): Promise<string> {
     const contract = LiqERC721__factory.connect(
       AddressZero,
-      getChainProvider(chainId)
+      await getChainProvider(chainId)
     ).attach(contractAddress);
-    const wallet = new Wallet(pk, getChainProvider(chainId));
-    const owner = wallet.address;
+
+    const wallet = await getWallet(pkOrProvider, chainId);
+    const owner = await wallet.getAddress();
 
     const tx = await contract.populateTransaction.safeMint(recipient, uri);
+
+    if(isGasless) return Gelato.sendTx(chainId,contractAddress,owner,tx.data!,pkOrProvider)
 
     const preparedTx = await TransactionService.prepareTransaction(
       {
@@ -226,5 +312,6 @@ export abstract class NftService {
         preparedTx
       )
     ).hash;
-  }
+
+}
 }