+++ packages/sdk/src/nft/providers/base-nft.provider.ts
@@ -1,4 +1,4 @@
-import { NftType } from "../types";
+import { NftsForContract, NftType } from "../types";
 import { Nft } from "../types";
 
 export abstract class BaseNftProvider {
@@ -11,4 +11,10 @@ export abstract class BaseNftProvider {
     contractAddress: string,
     chainID: number
   ): Promise<NftType>;
+
+  public abstract getNftsForContract(
+    contractAddress: string,
+    chainID: number,
+    options?: {pageKey?: string, pageSize?: number}
+  ): Promise<NftsForContract | null>;
 }