+++ packages/sdk/src/swap/debridge/debridgeSwapProvider.ts
@@ -3,18 +3,20 @@ import { getChainProvider } from "../../factory/chain-provider";
 import { TransactionService } from "../../transaction/transaction.service";
 import { DebridgeConfig } from "./config";
 import { FullSubmissionInfo, QuoteRequest, SwapRequest, SwapStatusRequest } from "./types";
-import { fetchGet, withInterval } from "../../common/utils";
+import { fetchGet, getWallet, withInterval } from "../../common/utils";
 import SignatureVerifier from "./abi/SignatureVerifier.json";
 import { AddressZero } from "@ethersproject/constants";
 import { ERC20, ERC20__factory } from "../../../typechain-types";
-import { formatEther, formatUnits, parseEther, parseUnits } from "ethers/lib/utils";
+import { formatUnits, parseEther, parseUnits } from "ethers/lib/utils";
 import BigNumber from "bignumber.js";
 import { TX_STATUS } from "../../transaction/constants/transaction-status";
+import { JsonRpcSigner, ExternalProvider } from "@ethersproject/providers";
+import { Gelato } from "../../gasless-providers/gelato";
 
 export abstract class DebridgeSwapProvider {
   private static readonly config = DebridgeConfig;
 
-  public static async swap(swapRequest: SwapRequest, pk: string) {
+  public static async swap(swapRequest: SwapRequest, pkOrProvider: string | ExternalProvider, isGasless: boolean) {
     // Validation
     const isSrcChainSupported = !!this.config.chains[swapRequest.srcChainId];
     const isDstChainSupported = !!this.config.chains[swapRequest.srcChainId];
@@ -25,15 +27,18 @@ export abstract class DebridgeSwapProvider {
     )
       throw Error("Swap not supported");
 
-    const chainProvider = getChainProvider(swapRequest.srcChainId);
-    const wallet = new Wallet(pk, chainProvider);
+    const isSwapFromNative = swapRequest.srcChainTokenIn !== AddressZero;
+
+    if(isGasless && isSwapFromNative) throw Error("Gasless not supported when swapping from native");
+
+    const wallet = await getWallet(pkOrProvider, swapRequest.srcChainId);
     let fromAmountInUnits: string;
 
-    if (swapRequest.srcChainTokenIn !== AddressZero) {
+    if (isSwapFromNative) {
       // Approve token
       const contract: ERC20 = ERC20__factory.connect(
         AddressZero,
-        chainProvider
+        await getChainProvider(swapRequest.srcChainId)
       ).attach(swapRequest.srcChainTokenIn);
       const decimals = await contract.decimals();
 
@@ -50,8 +55,8 @@ export abstract class DebridgeSwapProvider {
     }
 
     if (swapRequest.srcChainId === swapRequest.dstChainId)
-      return this.singleChainSwap(swapRequest, fromAmountInUnits!, wallet);
-    return this.crossChainSwap(swapRequest, fromAmountInUnits!, wallet);
+      return this.singleChainSwap(swapRequest, fromAmountInUnits!, wallet, isGasless, pkOrProvider);
+    return this.crossChainSwap(swapRequest, fromAmountInUnits!, wallet, isGasless, pkOrProvider);
   }
 
   public static async getQuote(quoteRequest: QuoteRequest) {
@@ -65,14 +70,13 @@ export abstract class DebridgeSwapProvider {
     )
       throw Error("Swap not supported");
 
-    const chainProvider = getChainProvider(quoteRequest.srcChainId);
     let fromAmountInUnits: string;
     let dstTokenDecimals: number;
 
     if (quoteRequest.srcChainTokenIn !== AddressZero) {
       const contract: ERC20 = ERC20__factory.connect(
         AddressZero,
-        chainProvider
+        await getChainProvider(quoteRequest.srcChainId)
       ).attach(quoteRequest.srcChainTokenIn);
       const decimals = await contract.decimals();
 
@@ -89,7 +93,7 @@ export abstract class DebridgeSwapProvider {
     if (quoteRequest.dstChainTokenOut !== AddressZero) {
       const contract: ERC20 = ERC20__factory.connect(
         AddressZero,
-        chainProvider
+        await getChainProvider(quoteRequest.dstChainId)
       ).attach(quoteRequest.dstChainTokenOut);
       dstTokenDecimals = await contract.decimals();
     } else {
@@ -132,7 +136,9 @@ export abstract class DebridgeSwapProvider {
   private static async singleChainSwap(
     swapRequest: SwapRequest,
     fromAmountInUnits: string,
-    wallet: Wallet
+    wallet: Wallet | JsonRpcSigner,
+    isGasless: boolean,
+    pkOrProvider: string | ExternalProvider
   ) {
     const {
       srcChainId: chainId,
@@ -150,13 +156,17 @@ export abstract class DebridgeSwapProvider {
       tokenOutRecipient,
     });
 
-    if (!response.tx) throw Error("Swap Tx Could not be created...Try again perhaps with different params");
+    const tx = response.tx;
+    if (!tx) throw Error("Swap Tx Could not be created...Try again perhaps with different params");
+
+    const from = await wallet.getAddress();
+    if(isGasless) return Gelato.sendTx(chainId,tx.to,from,tx.data!,pkOrProvider);
 
     const preparedTx = await TransactionService.prepareTransaction(
       {
-        from: wallet.address,
-        ...response.tx,
-        value: EthersBigNumber.from(response.tx.value)
+        from,
+        ...tx,
+        value: EthersBigNumber.from(tx.value)
       },
       chainId
     );
@@ -170,7 +180,7 @@ export abstract class DebridgeSwapProvider {
   private static async crossChainSwapQuote(
     quoteRequest: QuoteRequest,
     fromAmountInUnits: string,
-    decimals: number,
+    decimals: number
   ) {
     const {
       srcChainId,
@@ -197,7 +207,10 @@ export abstract class DebridgeSwapProvider {
   private static async crossChainSwap(
     swapRequest: SwapRequest,
     fromAmountInUnits: string,
-    wallet: Wallet
+    wallet: Wallet | JsonRpcSigner,
+    isGasless: boolean,
+    pkOrProvider: string | ExternalProvider
+
   ) {
     const {
       srcChainId,
@@ -215,13 +228,18 @@ export abstract class DebridgeSwapProvider {
       dstChainTokenOutRecipient,
     });
 
-    if (!response.tx) throw Error("Swap Tx Could not be created...Try again perhaps with different params");
+    const tx = response.tx;
+    if (!tx) throw Error("Swap Tx Could not be created...Try again perhaps with different params");
+
+    const from = await wallet.getAddress();
+
+    if(isGasless) return Gelato.sendTx(srcChainId,tx.to,from,tx.data!,pkOrProvider);
 
     const preparedTx = await TransactionService.prepareTransaction(
       {
-        from: wallet.address,
-        ...response.tx,
-        value: EthersBigNumber.from(response.tx.value)
+        from,
+        ...tx,
+        value: EthersBigNumber.from(tx.value)
       },
       srcChainId
     );
@@ -235,9 +253,9 @@ export abstract class DebridgeSwapProvider {
   private static async approveToken(
     contract: ERC20,
     amount: string,
-    wallet: Wallet
+    wallet: Wallet  | JsonRpcSigner
   ) {
-    if (await this.hasAllowance(contract, amount, wallet.address)) return true;
+    if (await this.hasAllowance(contract, amount, await wallet.getAddress())) return true;
 
     const tx = await contract.populateTransaction.approve(
       this.config.routerAddress,
@@ -311,7 +329,7 @@ export abstract class DebridgeSwapProvider {
       const signatureVerifier = new ethers.Contract(
         this.config.chains[srcChainId].signatureVerifier,
         SignatureVerifier.abi,
-        getChainProvider(srcChainId)
+        await getChainProvider(srcChainId)
       );
       const minConfirmations = await signatureVerifier.minConfirmations();
       const count = await this.getConfirmationsCount(sendTxHash);