+++ apps/web-demo/CHANGELOG.md
@@ -1,5 +1,21 @@
 # tkey-test
 
+## 0.1.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [f566694]
+  - @liquality/wallet-sdk@2.2.0-next.0
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [b464847]
+- Updated dependencies [f00c36f]
+- Updated dependencies
+  - @liquality/wallet-sdk@2.0.0
+
 ## 0.1.1-next.0
 
 ### Patch Changes